// src/ReportGenerator/model/data/TeacherJournal2.ts
export class TeacherJournal2 {
  teacher_name_c: string;
  department_short: string;
  journals_title: string;
  order_authors: string;
  address_authors: string;
  journals_type: string;
  journals_status_id: string;
  year_id: number;
  month_id: number;
  journals_class: string;
  other_authors: string;
  journals_name: string;
  project: string;
  volume: string|number;
  issue: string|number;
  page_s: string|number;
  page_e: string|number;
}